refactored generateBatchId function
