+++ src/providers/dataProvider/dataprovider-utils.ts
@@ -1,60 +1,9 @@
-import {
-  type DataProvider,
-  type UpdateParams,
-  type CreateResult
-} from 'ra-core'
+import { type UpdateParams, type CreateResult } from 'ra-core'
 import * as constants from '../../constants'
-import { isNumber } from '../../utils/number'
 import { DateTime } from 'luxon'
 import { type AuditType } from '../../utils/activity-types'
 import { isSameDate } from '../../utils/date'
 import { getUser } from '../authProvider'
-
-export const compareVersions = (v1: string, v2: string): number => {
-  const s1 = parseInt(v1.substring(1))
-  const s2 = parseInt(v2.substring(1))
-  if (isNaN(s1) || isNaN(s2)) return NaN
-  if (s1 < s2) {
-    return -1
-  } else if (s1 > s2) {
-    return 1
-  } else {
-    return 0
-  }
-}
-
-export const generateBatchId = async (
-  provider: DataProvider,
-  year: string
-): Promise<string> => {
-  if (!isNumber(year)) throw new TypeError('Year invalid')
-  const batches = await provider.getList(constants.R_BATCHES, {
-    sort: { field: 'id', order: 'ASC' },
-    pagination: { page: 1, perPage: 1000 },
-    filter: { yearOfReceipt: year }
-  })
-
-  if (batches.data.length === 0) {
-    return '00'
-  }
-
-  if (batches.data.length === 1) {
-    return '01'
-  }
-  const greatestBatch = batches.data.reduce((prev, current) =>
-    compareVersions(prev.batchNumber, current.batchNumber) === -1
-      ? current
-      : prev
-  )
-
-  return (parseInt(greatestBatch.batchNumber.substring(1)) + 1).toLocaleString(
-    'en-US',
-    {
-      minimumIntegerDigits: 2,
-      useGrouping: false
-    }
-  )
-}
 
 export const nowDate = (): string => {
   const isoDate = DateTime.utc().toISO()
