+++ src/providers/dataProvider/resource-callbacks/BatchLifeCycle.ts
@@ -1,8 +1,4 @@
-import {
-  type AuditFunctionType,
-  generateBatchId,
-  withCreatedByAt
-} from '../dataprovider-utils'
+import { type AuditFunctionType, withCreatedByAt } from '../dataprovider-utils'
 import {
   type CreateResult,
   type DataProvider,
@@ -11,6 +7,53 @@ import {
 import { AuditType } from '../../../utils/activity-types'
 import { R_BATCHES } from '../../../constants'
 import { type ResourceCallbacks } from 'react-admin'
+import { isNumber } from '../../../utils/number'
+
+const compareVersions = (v1: string, v2: string): number => {
+  const s1 = parseInt(v1.substring(1))
+  const s2 = parseInt(v2.substring(1))
+  if (isNaN(s1) || isNaN(s2)) return NaN
+  if (s1 < s2) {
+    return -1
+  } else if (s1 > s2) {
+    return 1
+  } else {
+    return 0
+  }
+}
+
+export const generateBatchId = async (
+  provider: DataProvider,
+  year: string
+): Promise<string> => {
+  if (!isNumber(year)) throw new TypeError('Year invalid')
+  const batches = await provider.getList(R_BATCHES, {
+    sort: { field: 'id', order: 'ASC' },
+    pagination: { page: 1, perPage: 1000 },
+    filter: { yearOfReceipt: year }
+  })
+
+  if (batches.data.length === 0) {
+    return '00'
+  }
+
+  if (batches.data.length === 1) {
+    return '01'
+  }
+  const greatestBatch = batches.data.reduce((prev, current) =>
+    compareVersions(prev.batchNumber, current.batchNumber) === -1
+      ? current
+      : prev
+  )
+
+  return (parseInt(greatestBatch.batchNumber.substring(1)) + 1).toLocaleString(
+    'en-US',
+    {
+      minimumIntegerDigits: 2,
+      useGrouping: false
+    }
+  )
+}
 
 export default (
   audit: AuditFunctionType,
